refactor: rename Token/Word to Word/Subword
