+++ ki-jetbrains/src/kotlin/protocol/Types.kt
@@ -111,14 +111,14 @@ sealed class SelectionMode {
 	@SerialName("LineFull")
 	object LineFull: SelectionMode()
 	@Serializable
-	@SerialName("Word")
-	object Word: SelectionMode()
+	@SerialName("Subword")
+	object Subword: SelectionMode()
 	@Serializable
 	@SerialName("WordFine")
 	object WordFine: SelectionMode()
 	@Serializable
-	@SerialName("Token")
-	object Token: SelectionMode()
+	@SerialName("Word")
+	object Word: SelectionMode()
 	@Serializable
 	@SerialName("Custom")
 	object Custom: SelectionMode()
