+++ ki-protocol-types/src/lib.rs
@@ -67,9 +67,9 @@ pub enum SelectionMode {
     Character,
     Line,
     LineFull,
-    Word,
+    Subword,
     WordFine,
-    Token,
+    Word,
     Custom,
     SyntaxNode,
     SyntaxNodeFine,
