+++ src/embed/app.rs
@@ -499,15 +499,8 @@ impl EmbeddedApp {
             crate::selection::SelectionMode::Line | crate::selection::SelectionMode::LineFull => {
                 ki_protocol_types::SelectionMode::Line
             }
-            crate::selection::SelectionMode::Word { skip_symbols: true } => {
-                ki_protocol_types::SelectionMode::Word
-            }
-            crate::selection::SelectionMode::Word {
-                skip_symbols: false,
-            } => ki_protocol_types::SelectionMode::WordFine,
-            crate::selection::SelectionMode::Token { .. } => {
-                ki_protocol_types::SelectionMode::Token
-            }
+            crate::selection::SelectionMode::Subword => ki_protocol_types::SelectionMode::Subword,
+            crate::selection::SelectionMode::Word { .. } => ki_protocol_types::SelectionMode::Word,
             crate::selection::SelectionMode::Custom => ki_protocol_types::SelectionMode::Custom,
             crate::selection::SelectionMode::Find { search } => {
                 ki_protocol_types::SelectionMode::Find {
