+++ src/selection.rs
@@ -420,8 +420,8 @@ impl SelectionSet {
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub(crate) enum SelectionMode {
     // Regex
-    Word { skip_symbols: bool },
-    Token,
+    Subword,
+    Word,
     Line,
     Character,
     Custom,
@@ -469,10 +469,8 @@ impl SelectionMode {
             SelectionMode::GitHunk(diff_mode) => format!("HUNK{}", diff_mode.display()).to_string(),
             SelectionMode::Mark => "MARK".to_string(),
             SelectionMode::LocalQuickfix { title } => title.to_string(),
-            SelectionMode::Word { skip_symbols } => {
-                format!("WORD{}", if *skip_symbols { "" } else { "*" })
-            }
-            SelectionMode::Token => "TOKEN".to_string(),
+            SelectionMode::Subword => "SUBWORD".to_string(),
+            SelectionMode::Word => "WORD".to_string(),
         }
     }
 
@@ -489,10 +487,8 @@ impl SelectionMode {
             cursor_direction,
         };
         Ok(match self {
-            SelectionMode::Word { skip_symbols } => {
-                Box::new(PositionBased(selection_mode::Word::new()))
-            }
-            SelectionMode::Token => Box::new(selection_mode::Token),
+            SelectionMode::Subword => Box::new(PositionBased(selection_mode::Subword::new())),
+            SelectionMode::Word => Box::new(selection_mode::Word),
             SelectionMode::Line => Box::new(PositionBased(selection_mode::LineTrimmed)),
             SelectionMode::LineFull => Box::new(PositionBased(selection_mode::LineFull::new())),
             SelectionMode::Character => Box::new(PositionBased(selection_mode::Character)),
@@ -532,8 +528,8 @@ impl SelectionMode {
     pub(crate) fn is_contiguous(&self) -> bool {
         matches!(
             self,
-            SelectionMode::Word { .. }
-                | SelectionMode::Token { .. }
+            SelectionMode::Subword { .. }
+                | SelectionMode::Word { .. }
                 | SelectionMode::Line
                 | SelectionMode::LineFull
                 | SelectionMode::Character
